Fix issue with modifying table rows
Change variables tableRowRemove in removeTableRow(e) and tableRow in
toggleRead(e).

If there were multiple rows that were added to the table and then a
row that was not the last row in the table was removed from the DOM
using removeTableRow(e), tableRowRemove would cause the wrong book
information to be removed from the DOM. Similarly, tableRow would cause
the wrong book object to be updated in the myLibrary array.